A&E attendances hourly moving averages
Starting with an initial total daily forecast for A& E patient
admissions (derived using Holt-Winters exponential smoothing), this is
then broken down by hour.

Quartiles for the forecast are also generated (i.e Lower, IQR and Upper
quartiles).

Then based on the actual live attendances per hour, the original
forecasted figure is adjusted depending on which quartile it is nearest
too.

Thus, a predicted value for the next hour and subsequent total for the
day is generated.